month names, cookies, etc
Changed login authentication from session variables to 10 year cookies not because it's safe but because martin thinks it's annoying to have to login once a day....

Changed revmonth and chargemonth from numbers to the month names,

Added 13 option to revmonth for customers that no longer accpets revisions.

Added a way to login through url again not because it's safe...

Removed console.logs and other minor changes.